refactor: Update tournament schema and match handling
- Changed the 'name' field in the groups schema to 'number' for better clarity.
- Added 'totalMatches', 'wins', and 'losses' fields to the groupPlayers schema to track player statistics.
- Updated TournamentManager to reflect changes in match retrieval and processing logic, ensuring consistency with the new schema.
- Refactored TournamentCard and TournamentMatchRow components to utilize updated match and group data, enhancing UI clarity and functionality.
