+++ backend/src/server.ts
@@ -18,7 +18,7 @@ import RatingManager from './managers/rating.manager'
 import SessionManager from './managers/session.manager'
 import SessionScheduler from './managers/session.scheduler'
 import TimeEntryManager from './managers/timeEntry.manager'
-import TournamentManager from './managers/tournament.manager'
+import TournamentManager from './managers/tournament.managers/tournament.manager'
 import TrackManager from './managers/track.manager'
 import UserManager from './managers/user.manager'
 
@@ -72,7 +72,7 @@ async function Connect(s: TypedSocket) {
   s.emit('all_time_entries', await TimeEntryManager.getAllTimeEntries())
   s.emit('all_matches', await MatchManager.getAllMatches())
   s.emit('all_rankings', await RatingManager.onGetRatings())
-  s.emit('all_tournaments', await TournamentManager.getAllTournaments())
+  s.emit('all_tournaments', await TournamentManager.getAll())
 
   s.on('disconnect', () =>
     console.debug(new Date().toISOString(), s.id, 'Disconnected')
