+++ drizzle/meta/_journal.json
@@ -78,6 +78,13 @@
       "when": 1767833425334,
       "tag": "0010_superb_lenny_balinger",
       "breakpoints": true
+    },
+    {
+      "idx": 11,
+      "version": "6",
+      "when": 1768062574602,
+      "tag": "0011_daffy_riptide",
+      "breakpoints": true
     }
   ]
 }
